Bug-fix: Views were not initialized to the view control's initial transform, so they were not viewing what they were supposed to view. Fixed by adding 'update()' to constructor.
Bug-fix: Views were not properly updated when the associated 'ViewControl' was updated. A temporary work-around was found by adding an anonymous 'DragListener' in 'Java3dView' to call 'update()' when needed. The problem was found in 'ViewModelView' where the original propery change listener was monitoring the 'tranform' property instead of 'transform' (typo!). Fixed typo and removed temporary work-around to solve the problem correctly.

Goal compliance fixes.