A number of problems with the render queue and the painter's algorithm have been fixed, although there are still some left. Debugging is a bit easier now since it is now possible to print the sorting proces to standard output (just uncomment all output statements).
A great number of cosmetic and code maturity fixes have been made

The difficult to understand methods 'order()', 'clip()' and 'compare()' now have some comments added to (hopefully) make them easier to read.

The method 'sortPolygonList()' has been added, which can sort a given list of 'RenderedPolygons' based on their z value. This method is only for debugging and should be deleted when all problems have been fixed.

The method 'sortQueue()' has been added, which sorts all polygons in a queue using the 'order()' method. This is a debugging method as well, but it's contents should be moved to 'getQueuedPolygons()' when problems have been fixed.